Pass record to Model constructor
Previously we had no explicit Model constructor, and we were preparing
the object by futzing with it in the delegating caster machinery. That
allowed a user to implement their own constructor, though since it
didn't receive any arguments it would have been of limited utility. Now
a user can receive the raw query result in their Model subclass
constructor (or __new__, which is actually what I want that's driving
this).